+++ screens/main/diet/RecipeDetailScreen.tsx
@@ -30,6 +30,7 @@ import useTodayProgressStore from "../../../stores/useTodayProgressStore";
 import useToastStore from "../../../stores/useToastStore";
 import * as haptics from "expo-haptics";
 import { trackMixpanelEvent } from "../../../services/mixpanel";
+import { analyticsEvents } from "../../../constants/analytics";
 type RecipeDetailScreenRouteProp = RouteProp<
   RootStackParamList,
   "RecipeDetailScreen"
@@ -91,7 +92,7 @@ const RecipeDetailScreen: React.FC<RecipeDetailScreenProps> = ({ route }) => {
         used: todayProgress.points.used + parseInt(recipe.points),
       },
     });
-    trackMixpanelEvent("recipe_added_to_diet", { recipe: recipe.title });
+    trackMixpanelEvent(analyticsEvents.dietRecipeAdded, { recipe: recipe.title });
     setTodayDiet([...todayDiet, payload]);
     addToDiet(user?.email, payload);
     navigation.goBack();